auth-service/requirements.txt
grpcio==1.60.0
grpcio-tools==1.60.0
passlib==1.7.4
PyJWT==2.8.0
mongoengine==0.24.2
pymongo==3.12.3
python-dotenv==1.0.0
protobuf==4.25.1